flip the nodes
