+++ gatsby-node.js
@@ -2,7 +2,13 @@
 /* eslint-env node */
 const path = require("path");
 
-exports.onCreateWebpackConfig = ({ loaders, actions, stage, getConfig }) => {
+exports.onCreateWebpackConfig = ({
+  loaders,
+  rules,
+  actions,
+  stage,
+  getConfig,
+}) => {
   const config = getConfig();
 
   /**
@@ -36,6 +42,37 @@ exports.onCreateWebpackConfig = ({ loaders, actions, stage, getConfig }) => {
       },
     ],
   });
+
+  /**
+   * Gatsby does not like that we use css modules. To fix this we need
+   * to change some of the webpack config around how we handle css.
+   * 😢 More info here: https://github.com/gatsbyjs/gatsby/issues/16129
+   */
+  const cssRule = {
+    ...rules.cssModules(),
+    test: rules.css().test,
+    include: /^((?!node_modules).)*$/,
+  };
+
+  /**
+   * Don't process css from npm packages as modules.
+   */
+  const libCssRule = {
+    ...rules.css(),
+    test: rules.css().test,
+    include: /node_modules/,
+  };
+
+  config.module.rules = [
+    ...config.module.rules.filter(rule => {
+      const areCssRules =
+        rule.oneOf && rule.oneOf.some(r => r.test.test("style.css"));
+
+      return !areCssRules;
+    }),
+    libCssRule,
+    cssRule,
+  ];
 
   // Situationally disable serverside rendering.
   if (stage.includes("html")) {
